cnidarium::proto
Module v1

Available on crate feature proto only.

Modules

key_value_response

Nested message and enum types in KeyValueResponse.

non_verifiable_key_value_request

Nested message and enum types in NonVerifiableKeyValueRequest.

non_verifiable_key_value_response

Nested message and enum types in NonVerifiableKeyValueResponse.

query_service_clientrpc

Generated client implementations.

query_service_serverrpc

Generated server implementations.

watch_response

Nested message and enum types in WatchResponse.

Structs

KeyValueRequest

Performs a key-value query against the JMT, either by key or by key hash.

KeyValueResponse

NonVerifiableKeyValueRequest

Performs a key-value query against the nonverifiable storage, using a byte-encoded key.

NonVerifiableKeyValueResponse

PrefixValueRequest

Performs a prefixed key-value query, by string prefix.

PrefixValueResponse

WatchRequest

Requests a stream of new key-value pairs that have been committed to the state.

WatchResponse

A key-value pair that has been committed to the state.