cnidarium
Trait StateWrite

pub trait StateWrite:
    StateRead
    + Send
    + Sync {
    // Required methods
    fn put_raw(&mut self, key: String, value: Vec<u8>);
    fn delete(&mut self, key: String);
    fn nonverifiable_put_raw(&mut self, key: Vec<u8>, value: Vec<u8>);
    fn nonverifiable_delete(&mut self, key: Vec<u8>);
    fn object_put<T: Clone + Any + Send + Sync>(
        &mut self,
        key: &'static str,
        value: T,
    );
    fn object_delete(&mut self, key: &'static str);
    fn object_merge(
        &mut self,
        objects: BTreeMap<&'static str, Option<Box<dyn Any + Send + Sync>>>,
    );
    fn record(&mut self, event: Event);
}

Write access to chain state.

Required Methods

fn put_raw(&mut self, key: String, value: Vec<u8>)

Puts raw bytes into the verifiable key-value store with the given key.

fn delete(&mut self, key: String)

Delete a key from the verifiable key-value store.

fn nonverifiable_put_raw(&mut self, key: Vec<u8>, value: Vec<u8>)

Puts raw bytes into the non-verifiable key-value store with the given key.

fn nonverifiable_delete(&mut self, key: Vec<u8>)

Delete a key from non-verifiable key-value storage.

fn object_put<T: Clone + Any + Send + Sync>( &mut self, key: &'static str, value: T, )

Puts an object into the ephemeral object store with the given key.

Panics

If the object is already present in the store, but its type is not the same as the type of value.

fn object_delete(&mut self, key: &'static str)

Deletes a key from the ephemeral object store.

fn object_merge( &mut self, objects: BTreeMap<&'static str, Option<Box<dyn Any + Send + Sync>>>, )

Merge a set of object changes into this StateWrite.

Unlike object_put, this avoids re-boxing values and messing up the downcasting.

fn record(&mut self, event: Event)

Record that an ABCI event occurred while building up this set of state changes.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.